eliminate compiler warnings

+++ coders/wpg.c
@@ -600,8 +600,8 @@ static int UnpackWPG2Raster(Image *image,int bpp)
     return(-2);
   (void) memset(BImgBuff,0,ldblk*sizeof(*BImgBuff));
 
-  while( y< image->rows)
-    {
+  while (y < (ssize_t) image->rows)
+  {
       bbuf=ReadBlobByte(image);
 
       switch(bbuf)
@@ -658,7 +658,7 @@ static int UnpackWPG2Raster(Image *image,int bpp)
             /* duplicate the previous row RunCount x */
             for(i=0;i<=RunCount;i++)
               {
-                if (InsertRow(BImgBuff,(ssize_t) (image->rows > y ? y : image->rows-1),image,bpp) == MagickFalse)
+                if (InsertRow(BImgBuff,(ssize_t) image->rows > y ? y : (ssize_t) image->rows-1,image,bpp) == MagickFalse)
                   {
                     BImgBuff=(unsigned char *) RelinquishMagickMemory(BImgBuff);
                     return(-3);
